feat: add heavy equipment detection and support fp32

+++ src/config.py
@@ -1,6 +1,6 @@
 import os
 
-DEFAULT_PRECISION = "fp16"
+DEFAULT_PRECISION = "fp32"
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 STATIC_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), "main/static"))
